fix(global_weather.csv): convert sunrise, sunset, timezone_hour to time with format(hh:mm:ss)

+++ scripts/convert_date.py
@@ -4,4 +4,4 @@ def to_local_hour(date_in_second, timezone_in_second):
     return datetime.fromtimestamp(date_in_second + timezone_in_second).time()
 
 def timezone_to_hour(timezone_in_second):
-    return timezone_in_second / 3600
+    return int(timezone_in_second / 3600)
